added guard for front row disable recipe

+++ recipes/disable_front_row.rb
@@ -1,4 +1,5 @@
 execute "move front row out of the way" do
   command "mv /System/Library/CoreServices/Front\\ Row.app /System/Library/CoreServices/Front_Row_disabled.app"
   not_if { File.exists?("/System/Library/CoreServices/Front_Row_disabled.app") }
+  only_if { File.exists?("/System/Library/CoreServices/Front\\ Row.app") }
 end
